requirements.txt
flask
gspread
google-auth
pandas
pymupdf
google-generativeai
python-dotenv
